pyproject.toml
[tool.poetry]
name = "sa-senai-bombeiros_backend"
version = "0.1.0"
description = ""
authors = ["uBrunoow <brunowerner7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.7.2"
django = "4.2.6"
django-jazzmin = "2.6.0"
pillow = "10.1.0"
prettyconf = "2.2.1"
sqlparse = "0.4.4"
typing-extensions = "4.8.0"
aiohttp = "3.8.6"
aiosignal = "1.3.1"
anyio = "4.0.0"
async-timeout = "4.0.3"
attrs = "23.1.0"
boto3 = "1.28.74"
botocore = "1.31.74"
cerberus = "1.3.5"
certifi = "2023.7.22"
charset-normalizer = "3.3.1"
click = "8.1.7"
cssselect2 = "0.7.0"
django-embed-video = "1.4.9"
django-extensions = "3.2.3"
django-faicon = "0.1.0"
django-filer = "3.1.0"
django-material-icon-widget = "0.1.2"
django-polymorphic = "3.1.0"
django-storages = "1.14.2"
django-user-sessions = "2.0.0"
easy-thumbnails = "2.8.5"
exceptiongroup = "1.1.3"
fontawesomefree = "6.4.2"
frozenlist = "1.4.0"
geoip2 = "4.7.0"
h11 = "0.14.0"
httpcore = "0.18.0"
httpx = "0.25.0"
idna = "3.4"
install = "1.3.5"
jmespath = "1.0.1"
kolo = "2.15.1"
lxml = "4.9.3"
maxminddb = "2.4.0"
multidict = "6.0.4"
pyparsing = "3.1.1"
python-dateutil = "2.8.2"
pyyaml = "6.0.1"
reportlab = "4.0.6"
requests = "2.31.0"
s3transfer = "0.7.0"
simplejson = "3.19.2"
six = "1.16.0"
sniffio = "1.3.0"
svglib = "1.5.1"
tinycss2 = "1.2.1"
tomli = "2.0.1"
toolz = "0.12.0"
ulid-py = "1.1.0"
urllib3 = "2.0.7"
webencodings = "0.5.1"
yarl = "1.9.2"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.0"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.1"
django-rest-passwordreset = "^1.3.0"
django-filter = "^23.5"
django-username-email = "^2.5.4"
pytest = "^8.0.2"


[tool.poetry.group.dev-dependencies.dependencies]
pre-commit = "^3.6.2"
isort = "^5.13.2"
flake8 = "^7.0.0"
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
